fix(lint): split service.rs test/trigger family under the 600-line gate
service_trigger_tests.rs (742) and service_tests.rs (661) had grown past the
600-line pre-push gate; service_trigger.rs (650) was also over. Splits by
concern, following the repo's sibling-file convention: ceiling-validation
tests, snooze/skip_runs tests, ANSI-log-tail tests, and routine-flag CRUD each
move to a new sibling file. No behavior or test-body changes.
