+++ src/routines/service_tests.rs
@@ -446,25 +446,6 @@ fn svc_update_rejects_zero_durations() {
 }
 
 #[test]
-fn svc_create_rejects_ttl_above_cron_ceiling() {
-    let _home = TempHome::set();
-    // A `*/5 * * * *` routine has a ttl ceiling of min(3600, 300) = 300s. An explicit 1800 would be
-    // silently clamped to 300, so it is rejected with `BadRequest` up front (#468).
-    let store = new_store();
-    let result = svc_create(
-        &store,
-        CreateRoutineRequest {
-            model: None,
-            goal: None,
-            schedule: "*/5 * * * *".into(),
-            ttl_secs: Some(1800),
-            ..valid_create_request()
-        },
-    );
-    assert!(matches!(result, Err(AppError::BadRequest(_))));
-}
-
-#[test]
 fn svc_create_rejects_agent_config_without_prompt_placeholder() {
     // A parseable, registered agent whose `args` carry no `{prompt}`/`{prompt_file}` placeholder
     // would launch with no task. It is rejected at create time with `BadRequest` (#322), not
@@ -495,106 +476,6 @@ fn svc_create_rejects_agent_config_without_prompt_placeholder() {
     }
 
     std::fs::remove_file(&cfg).unwrap();
-}
-
-#[test]
-fn svc_create_rejects_max_runtime_above_cron_ceiling() {
-    let _home = TempHome::set();
-    // Mirror of the ttl ceiling for the watchdog bound (#468).
-    let store = new_store();
-    let result = svc_create(
-        &store,
-        CreateRoutineRequest {
-            model: None,
-            goal: None,
-            schedule: "*/5 * * * *".into(),
-            max_runtime_secs: Some(1800),
-            ..valid_create_request()
-        },
-    );
-    assert!(matches!(result, Err(AppError::BadRequest(_))));
-}
-
-#[test]
-fn svc_create_accepts_secs_at_cron_ceiling() {
-    let _home = TempHome::set();
-    // A value equal to the cron-derived ceiling (`*/5` -> 300s) is in force, not clamped, so it
-    // passes `reject_over_ceiling` (covering the `secs <= ceiling` arm for both fields). A
-    // duplicate-slug routine pre-seeded in the store makes the create fail *after* that check with a
-    // `Conflict`, so the assertion proves the ceiling check did not reject — without performing any
-    // crontab/disk mutation.
-    let store = store_with(vec![make_routine(
-        "at-ceiling-dupe",
-        "At Ceiling ZZZ",
-        1,
-        1,
-    )]);
-    let result = svc_create(
-        &store,
-        CreateRoutineRequest {
-            model: None,
-            goal: None,
-            schedule: "*/5 * * * *".into(),
-            // Same slug as the pre-seeded routine.
-            title: "  at   ceiling ZZZ ".into(),
-            ttl_secs: Some(300),
-            max_runtime_secs: Some(300),
-            ..valid_create_request()
-        },
-    );
-    assert!(matches!(result, Err(AppError::Conflict(_))));
-}
-
-#[test]
-fn svc_update_rejects_ttl_above_current_schedule_ceiling() {
-    let _home = TempHome::set();
-    // No schedule supplied: the ceiling derives from the routine's *current* `*/5` schedule, so a
-    // 1800s ttl exceeds the 300s ceiling and is rejected without mutating the store (#468).
-    let store = store_with(vec![Routine {
-        schedule: "*/5 * * * *".to_string(),
-        ..make_routine("upd-ttl-ceiling", "Keep Ceiling", 1, 1)
-    }]);
-    let result = svc_update(
-        &store,
-        "upd-ttl-ceiling",
-        UpdateRoutineRequest {
-            model: None,
-            goal: None,
-            ttl_secs: Some(1800),
-            ..empty_update_request()
-        },
-    );
-    assert!(matches!(result, Err(AppError::BadRequest(_))));
-    // The store value is untouched by the rejected update.
-    assert_eq!(
-        store
-            .lock()
-            .unwrap()
-            .get("upd-ttl-ceiling")
-            .unwrap()
-            .ttl_secs,
-        None
-    );
-}
-
-#[test]
-fn svc_update_rejects_secs_above_new_schedule_ceiling() {
-    let _home = TempHome::set();
-    // A supplied schedule is the *effective* schedule for the ceiling: tightening a `@daily` routine
-    // to `*/5` while setting max_runtime 1800 exceeds the new 300s ceiling and is rejected (#468).
-    let store = store_with(vec![make_routine("upd-new-sched", "Keep New Sched", 1, 1)]);
-    let result = svc_update(
-        &store,
-        "upd-new-sched",
-        UpdateRoutineRequest {
-            model: None,
-            goal: None,
-            schedule: Some("*/5 * * * *".into()),
-            max_runtime_secs: Some(1800),
-            ..empty_update_request()
-        },
-    );
-    assert!(matches!(result, Err(AppError::BadRequest(_))));
 }
 
 #[test]
